+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/register/dao/HomeTracingRegisterDao.kt
@@ -22,6 +22,7 @@ import javax.inject.Singleton
 import org.hl7.fhir.r4.model.Coding
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
+import org.smartregister.fhircore.engine.data.local.tracing.TracingRepository
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 
 @Singleton
@@ -29,10 +30,18 @@ class HomeTracingRegisterDao
 @Inject
 constructor(
   fhirEngine: FhirEngine,
+  tracingRepository: TracingRepository,
   defaultRepository: DefaultRepository,
   configurationRegistry: ConfigurationRegistry,
   dispatcherProvider: DefaultDispatcherProvider
-) : TracingRegisterDao(fhirEngine, defaultRepository, configurationRegistry, dispatcherProvider) {
+) :
+  TracingRegisterDao(
+    fhirEngine,
+    defaultRepository,
+    tracingRepository,
+    configurationRegistry,
+    dispatcherProvider
+  ) {
 
   override val tracingCoding: Coding = Coding("https://d-tree.org", "home-tracing", "Home Tracing")
 }
