+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/register/dao/TracingRegisterDao.kt
@@ -22,18 +22,14 @@ import com.google.android.fhir.FhirEngine
 import com.google.android.fhir.get
 import com.google.android.fhir.logicalId
 import com.google.android.fhir.search.Operation
-import com.google.android.fhir.search.Order
 import com.google.android.fhir.search.Search
 import com.google.android.fhir.search.has
 import com.google.android.fhir.search.search
 import java.util.Date
 import org.hl7.fhir.r4.model.CarePlan
-import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.DateTimeType
-import org.hl7.fhir.r4.model.Encounter
 import org.hl7.fhir.r4.model.Identifier
-import org.hl7.fhir.r4.model.ListResource
 import org.hl7.fhir.r4.model.Observation
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Practitioner
@@ -45,9 +41,9 @@ import org.smartregister.fhircore.engine.configuration.app.AppConfigClassificati
 import org.smartregister.fhircore.engine.configuration.app.ApplicationConfiguration
 import org.smartregister.fhircore.engine.data.domain.Guardian
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
+import org.smartregister.fhircore.engine.data.local.tracing.TracingRepository
 import org.smartregister.fhircore.engine.domain.model.ProfileData
 import org.smartregister.fhircore.engine.domain.model.RegisterData
-import org.smartregister.fhircore.engine.domain.model.TracingAttempt
 import org.smartregister.fhircore.engine.domain.repository.RegisterDao
 import org.smartregister.fhircore.engine.domain.util.PaginationConstant
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
@@ -67,8 +63,9 @@ abstract class TracingRegisterDao
 constructor(
   open val fhirEngine: FhirEngine,
   val defaultRepository: DefaultRepository,
+  private val tracingRepository: TracingRepository,
   val configurationRegistry: ConfigurationRegistry,
-  val dispatcherProvider: DefaultDispatcherProvider
+  val dispatcherProvider: DefaultDispatcherProvider,
 ) : RegisterDao {
   protected abstract val tracingCoding: Coding
 
@@ -128,28 +125,8 @@ constructor(
           )
           .patientTypeFilterTagViaMetaCodingSystem
       val tasks = validTasks(patient)
-      var currentAttempt: TracingAttempt? = null
-      val list =
-        fhirEngine.search<ListResource> {
-          filter(ListResource.SUBJECT, { value = patient.referenceValue() })
-          filter(ListResource.STATUS, { value = of(ListResource.ListStatus.CURRENT.toCode()) })
-          sort(ListResource.DATE, Order.ASCENDING)
-          count = 1
-          from = 0
-        }
-      val first = list.firstOrNull()
-      currentAttempt =
-        if (first != null) {
-          toTracingAttempt(first)
-        } else {
-          TracingAttempt(
-            historyId = null,
-            lastAttempt = null,
-            numberOfAttempts = 0,
-            outcome = "",
-            reasons = validTasks(it).mapNotNull { task -> task.reasonCode?.codingFirstRep?.display }
-          )
-        }
+
+      val attempt = tracingRepository.getTracingAttempt(patient)
 
       ProfileData.TracingProfileData(
         logicalId = patient.logicalId,
@@ -170,69 +147,13 @@ constructor(
         conditions = patient.activeConditions(),
         guardians = patient.guardians(),
         practitioners = patient.practitioners(),
-        currentAttempt = currentAttempt,
-        lastAttempt = currentAttempt,
-        numberOfAttempts = 0
+        currentAttempt =
+          attempt.copy(
+            reasons =
+              validTasks(patient).mapNotNull { task -> task.reasonCode?.codingFirstRep?.display }
+          ),
       )
     }
-  }
-
-  private suspend fun toTracingAttempt(list: ListResource): TracingAttempt {
-    val tasks = mutableListOf<Task>()
-    val encounters = mutableListOf<Encounter>()
-    var lastAttempt: Encounter? = null
-    val reasons = mutableListOf<String>()
-    var outcome = ""
-    list.entry.forEach { entry ->
-      try {
-        val ref = entry.item
-        val el = ref.reference.split("/")
-        val resource = fhirEngine.get(ResourceType.fromCode(el[0]), el[1])
-
-        if (resource is Task) {
-          resource.reasonCode?.codingFirstRep?.display?.let { reasons.add(it) }
-          tasks.add(resource)
-        } else if (resource is Encounter) {
-
-          if (lastAttempt == null) {
-            lastAttempt = resource
-          } else if (lastAttempt?.period?.start != null) {
-            if (resource.period.start.after(lastAttempt?.period?.start)) {
-              lastAttempt = resource
-            }
-          }
-          encounters.add(resource)
-        }
-      } catch (e: Exception) {
-        e.printStackTrace()
-      }
-    }
-    if (lastAttempt != null) {
-      val obs =
-        fhirEngine.search<Observation> {
-          filter(Observation.ENCOUNTER, { value = lastAttempt?.referenceValue() })
-          filter(
-            Observation.CODE,
-            {
-              value =
-                of(CodeableConcept(Coding("https://d-tree.org", "tracing-outcome-conducted", "")))
-            },
-            {
-              value =
-                of(CodeableConcept(Coding("https://d-tree.org", "tracing-outcome-unconducted", "")))
-            },
-            operation = Operation.OR
-          )
-        }
-      obs.firstOrNull()?.let { outcome = it.code.text ?: "" }
-    }
-    return TracingAttempt(
-      numberOfAttempts = encounters.size,
-      lastAttempt = lastAttempt?.period?.start,
-      outcome = outcome,
-      reasons = reasons,
-      historyId = list.logicalId
-    )
   }
 
   suspend fun Patient.activeCarePlans() =
