+++ android/engine/src/main/java/org/smartregister/fhircore/engine/domain/model/ProfileData.kt
@@ -111,9 +111,7 @@ sealed class ProfileData(open val logicalId: String, open val name: String) {
     val address: String,
     val chwAssigned: Reference,
     val healthStatus: HealthStatus,
-    val numberOfAttempts: Int,
     val currentAttempt: TracingAttempt?,
-    val lastAttempt: TracingAttempt?,
     val age: String = birthdate.toAgeDisplay(),
     val gender: Enumerations.AdministrativeGender,
     val addressDistrict: String = "",
