+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/shared/models/PatientProfileViewData.kt
@@ -97,11 +97,9 @@ sealed class ProfileViewData(
     val isHomeTracing: Boolean = false,
     val sex: String = "",
     val age: String = "",
-    val attempts: Int = 0,
     val dueDate: String = "",
     val identifierKey: String = "",
     val currentAttempt: TracingAttempt? = null,
-    val lastAttempt: TracingAttempt? = null,
     val showIdentifierInProfile: Boolean = false,
     val addressDistrict: String = "",
     val addressTracingCatchment: String = "",
