+++ android/quest/src/main/java/org/smartregister/fhircore/quest/util/mappers/ProfileViewDataMapper.kt
@@ -181,9 +181,7 @@ class ProfileViewDataMapper @Inject constructor(@ApplicationContext val context:
           sex = inputModel.gender.translateGender(context),
           age = inputModel.age,
           isHomeTracing = inputModel.tasks.firstOrNull { x -> x.isHomeTracingTask() } != null,
-          attempts = inputModel.numberOfAttempts,
           currentAttempt = inputModel.currentAttempt,
-          lastAttempt = inputModel.lastAttempt,
           dueDate = "",
           identifierKey = inputModel.healthStatus.retrieveDisplayIdentifierKey(),
           showIdentifierInProfile = false,
